[feat] 댓글 삭제 추가

+++ src/main/java/com/cmchackathon/domain/comment/service/CommentService.java
@@ -23,6 +23,10 @@ public class CommentService {
             throw new BusinessException(ErrorCode.FORBIDDEN);
         }
 
+        if (comment.getDeletedAt() != null) {
+            throw new BusinessException(ErrorCode.COMMENT_NOT_FOUND);
+        }
+
         comment.delete();
     }
 }
